moved gateway event tests
